Use explicit double to float conversions to avoid warnings
Such conversion warnings are enabled by default on MSVC.

On gcc/clang -Wconversion is needed for such warnings, and it is still
not enabled by default in "configure" because there are ~100 lines that
cause these warnings (some are in include files, so the total number of
warnings is ~1700).

+++ link-grammar/utilities.c
@@ -768,7 +768,7 @@ bool strtodC(const char *s, float *r)
 
 	if ('\0' != *err) return false; /* *r unaffected */
 
-	*r = val;
+	*r = (float)val;
 	return true;
 }
